requirements.txt
# Requirements automatically generated by pigar.
# https://github.com/damnever/pigar

# api/admin.py: 1
# api/apps.py: 1
# api/migrations/0001_initial.py: 3,4,5
# api/models.py: 7,8
# api/schema.py: 3
# api/tests.py: 1
# manage.py: 8
# minesweeper_unchained/urls.py: 2,3,5
# minesweeper_unchained/wsgi.py: 12
Django == 1.11

# api/schema_relay.py: 4
django_filter == 1.1.0

# api/schema.py: 1
# api/schema_relay.py: 3
# minesweeper_unchained/schema.py: 1
graphene == 2.1.3

# api/schema.py: 2
# api/schema_relay.py: 5,6
# minesweeper_unchained/urls.py: 4
graphene_django == 2.1.0

# api/schema_relay.py: 7
graphql_relay == 0.4.5

# api/models.py: 9
numpy == 1.15.1

# minesweeper_unchained/settings.py: 14
python-dotenv == 0.9.1

# minesweeper_unchained/settings.py: 15
dj-database-url == 0.5.0

psycopg2-binary == 2.7.5
gunicorn == 19.9.0
whitenoise == 4.0
django-livereload-server == 0.3.2
django-cors-headers == 2.4.0
djangorestframework == 3.8.2
graphene-django-subscriptions == 0.0.4
